requirements.txt
flask
pandas
numpy
scikit-learn
nltk
emoji
sentence-transformers
transformers
torch
python-dotenv
